+++ app/Http/Controllers/ClassificacaoController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\Classificacao\StoreClassificacaoRequest;
 use App\Http\Requests\Classificacao\UpdateClassificacaoRequest;
 use App\Models\Classificacao;
 use App\Models\Patrimonio;
+use App\Models\Subgrupo;
 use Illuminate\Http\Request;
 
 class ClassificacaoController extends Controller
@@ -43,8 +44,14 @@ class ClassificacaoController extends Controller
     public function delete($classificacao_id)
     {
         $classificacao = Classificacao::find($classificacao_id);
+
+        if(!$classificacao->subgrupos()->exists()){
             $classificacao->delete();
+
             return redirect(route('classificacao.index'))->with('success', 'Classificação Removida com Sucesso!');
+        }else{
+            return redirect()->route('classificacao.index')->with('fail', 'Não é possivel remover esta classificação, há subgrupos vinculados a ela!');
+        }
     }
 
     public function search(Request $request)
